remove double client side quoting
