+++ src/nft/hooks/useDerivedPayWithAnyTokenSwapInfo.ts
@@ -1,6 +1,6 @@
 import { Currency, CurrencyAmount, NativeCurrency, Percent, Token, TradeType } from '@uniswap/sdk-core'
 import useAutoSlippageTolerance from 'hooks/useAutoSlippageTolerance'
-import { useBestTrade } from 'hooks/useBestTrade'
+import { useDebouncedTrade } from 'hooks/useDebouncedTrade'
 import { useMemo } from 'react'
 import { ClassicTrade, RouterPreference, TradeState } from 'state/routing/types'
 import { isClassicTrade } from 'state/routing/utils'
@@ -14,7 +14,7 @@ export default function useDerivedPayWithAnyTokenSwapInfo(
   maximumAmountIn?: CurrencyAmount<Token>
   allowedSlippage: Percent
 } {
-  const { state, trade } = useBestTrade(
+  const { state, trade } = useDebouncedTrade(
     TradeType.EXACT_OUTPUT,
     parsedOutputAmount,
     inputCurrency ?? undefined,
